backend/requirement.txt
aioredis==1.1.0
asgi-redis==0.14.1
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.6.1
Automat==0.6.0
autopep8==1.3.5
certifi==2018.4.16
cffi==1.11.5
channels==2.1.2
channels-redis==2.2.1
chardet==3.0.4
colorama==0.3.7
constantly==15.1.0
contextlib2==0.5.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
daphne==2.2.0
dj-database-url==0.4.1
Django==2.0.7
django-cors-headers==2.2.0
django-idmap==1.0.3
django-rest-auth==0.9.3
django-rest-knox==3.1.4
django-vanilla-views==1.0.4
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
flake8==3.5.0
hiredis==0.2.0
honcho==0.7.1
huey==1.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
IPy==0.83
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
msgpack==0.5.6
msgpack-python==0.4.8
otree==2.1.9
otree-boto2-shim==0.3.2
otree-core==0.0.0b1
otree-save-the-change==1.1.3
pbr==1.10.0
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2==2.7.5
py==1.4.31
pycodestyle==2.4.0
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
pytest==2.9.2
pytest-django==3.0.0
python-decouple==3.1
python-redis-lock==3.2.0
pytz==2017.3
raven==5.25.0
redis==2.10.5
requests==2.11.1
schema==0.6.2
six==1.10.0
termcolor==1.1.0
Twisted==18.4.0
txaio==2.10.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
whitenoise==3.2.1
ws4py==0.3.5
XlsxWriter==0.9.3
zope.interface==4.5.0
